third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-async/async-await/async-pause-on-exception-expected.txt

CONSOLE ERROR: line 37: Uncaught (in promise) Error: uncaught
CONSOLE ERROR: line 37: Uncaught (in promise) Error: uncaught
Tests that pause on promise rejection works.

=== Pausing only on uncaught exceptions ===
Set timer for test function.
Call stack:
    0) testFunction (async-pause-on-exception.html:37)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>
Paused on promise rejectionError: uncaught

=== Pausing on all exceptions ===
Set timer for test function.
Call stack:
    0) testFunction (async-pause-on-exception.html:36)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>
Paused on promise rejectionError: caught
Call stack:
    0) testFunction (async-pause-on-exception.html:37)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>
Paused on promise rejectionError: uncaught
